feat: migrate schema to UnixNano timestamps and OTel v1.39
Schema changes:
- DateTime → ulong nanoseconds (start_time_unix_nano, end_time_unix_nano)
- duration_ms → duration_ns
- decimal cost → double cost
- genai_* → gen_ai_* (OTel naming)
- Remove QueryParquet functionality

Test updates:
- SpanBuilder uses TimeProvider.System
- Assertions updated for new types
- Remove Parquet query tests (functionality removed)

+++ src/qyl.mcp/Tools/ReplayTools.cs
@@ -279,7 +279,7 @@ internal sealed record SessionSummary(
     [property: JsonPropertyName("total_output_tokens")]
     long TotalOutputTokens,
     [property: JsonPropertyName("total_cost_usd")]
-    decimal TotalCostUsd,
+    double TotalCostUsd,
     [property: JsonPropertyName("start_time")]
     string StartTime);
 
@@ -311,7 +311,7 @@ internal sealed record SpanDto(
     [property: JsonPropertyName("gen_ai_output_tokens")]
     long? GenAiOutputTokens,
     [property: JsonPropertyName("gen_ai_cost_usd")]
-    decimal? GenAiCostUsd);
+    double? GenAiCostUsd);
 
 internal sealed record TraceResponse(
     [property: JsonPropertyName("trace_id")]
